requirements.txt
websocket-client
pytest